Add files via upload
This is from a project from my Web Development Class in Fall of 2023. We were tasked with creating an online auction system for an elementary school. My part of the project, part 3 was to allow users to record bid information and contact winners. I also created very simple password protection with protect.php. When the password "Auction" was entered successfully, users were directed to a different home page. Users could then print off bidder forms, enter new bidder information, view and update bidder information, view and update lot data, as well as view every single bidder in the auction, whether they have bid on an item or not. A more detailed description of the assignment can be found under "CIS FINAL PROJECT" pdf, and the section I worked on was part 3.